02_cv-deep-learning/requirements.txt
torch
torchvision
mlflow
fastapi
uvicorn[standard]
pydantic
Pillow
numpy
pytest